Uploaded "Christopher Alexander.pdf", which is an edited version of https://www.patternlanguage.com/archive/ieee.html
C.A. talks about living structure, and implores the Software community to take on the responsibility of putting living structure back into the world. There are MANY parallels between this and what we are trying to do here, such that you might say that we are attempting to answer his call to that cause.